Create api-server-v1 to maintain backward compatibility
This is a temporary measure until V2 is ready to be widely adopted. This also restores the proper behavior for the vscode api (since the vscode client hasn’t been revved to support the new entry points). We can throw away the api-cache once we fully migrate to V2 / fully halt usage of V1.